Proyectos de la Comunidad/AsistenciaCIA/Lista-de-asistencia-con-reconocimiento-facial-usando-Python/requirements.txt
click==8.1.7
colorama==0.4.6
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
mysql-connector-python== 8.1.0
numpy==1.25.2
opencv-contrib-python==4.7.0.72
Pillow==10.0.1
pip==23.2.1
protobuf==4.21.12
reportlab==4.0.4
setuptools==68.2.2